Plugin CD: Fix problem with relative import
Without this commit applied, loading the panel fails because module importer cannot be imported.
